requirements.txt
youtube-transcript-api
langchain-community
faiss-cpu
tiktoken
python-dotenv
langchain-google-genai
langchain-huggingface
langchain-core



